fix: spring-boot version upgraded to 4.0.5, ProductRepository Abstract type changed from Long to Integer.

+++ backend/optcg/src/main/java/ual/dra/optcg/repository/ProductRepository.java
@@ -8,7 +8,7 @@ import ual.dra.optcg.entity.Product;
 import java.util.List;
 
 @RepositoryRestResource(path = "products")
-public interface ProductRepository extends CrudRepository<Product, Long> {
+public interface ProductRepository extends CrudRepository<Product, Integer> {
 
     public List<Product> findByCardName(String name);
